backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=13.0.0
google-cloud-aiplatform==1.38.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic>=2.9.0
python-multipart==0.0.6
google-genai==1.67.0
pydub==0.25.1
numpy==1.26.0
